Add smallest_denomination for Bitcoin (BTC)
The smallest value that the Bitcoin network supports sending is the
satoshi (sometimes abbreviated sat), one hundred-millionth
(0.000 000 01) of a bitcoin. In other words, the network does not
support sending fractions of a satoshi. Since it is a hard limit, it
seems natural to use it as a unit, though it currently has very little
value.
https://en.bitcoin.it/wiki/Help:FAQ#What_do_I_call_the_various_denominations_of_bitcoin.3F