Rewrite allocate/split
This commit introduces a more effective and uniform solution to
the #allocate and #split methods. These are now synonyms accepting
either array for allocations or a number for even splits. The solution
does not differ between the two approaches.

It also adds a couple of tests to make sure it works as expected, since
there were some inconsistencies with the former solution.