Respect symbol_first i18n setting when formatting amount
Previously only the separators were localized, but the symbol position
remained unchanged and based on the currency.
In most cases, when you localize an amount, the currency setting is less
important than the user preference (i.e. through locale).

Although this is not a rails default currency i18n setting, it would be
very helpful if it was respected correctly.

The fallback remains to the currency setting, so this change is
backwards compatible unless the "number.format.currency.symbol_first"
key is set.