Support systems in which subunit_to_unit ratio is always 100
Some systems are set up in a way where the assume that #cents
will always return 100 times the value displayed by #format.
This is only the case for currencies with subunit_to_unit ratio 100.

In order to cater to the needs of those systems without altering
the behavior of the gem we introduce a convience method which
anihilates the effect of subunit_to_unit ratio.