Add rubocop
I fixed the low hanging fruits without breaking the specs (at least
locally).

Rules that could not be fixed that easy were added as exception to the
`.rubocop.yml` configuration file.

Some of those exceptions might be fixable with some code recatoring,
which might be a good point to start with future PRs.