Fix CHF format regression introduced in v6.14 (#967)
* Remove symbol_with_space formatting specs

symbol_with_spec is not actually implmented in the codebase.
It looks like a leftover from initial commits of #943.

* Change format config for CHF currency

Looks like #943 introduced the format config only on currencies
that have `symbol_first` set to `false`. CHF is not one of them
and needs a different format.

This configuration is currently changing how CHF is displayed.

From:

	Money.new(1234, 'CHF').format # => "CHF12.34"

To:

	Money.new(1234, 'CHF').format # => "CHF 12.34"

This commit restores the old formatting.