Fix failing spec due to `symbol_position` defaults
Address the unexpected behavior where `symbol_position: :before` has a
different default compared to `symbol_position: :after`. The latter
defaults to adding a space, which is expected based on the
implementation in `determine_format_from_formatting_rules`.

This commit does not alter the current behavior but fixes the failing
spec to reflect the correct default behavior as per the implementation.
Changing the default would introduce breaking changes, thus this commit
only aims to make the spec pass.

Note: Further investigation into the different default behaviors for
`:before` and `:after` should be tracked in a separate issue.

Fix #1075