Allow .with_rounding_mode blocks nesting
So that it won't return to the global value in case if someone will
modify down in the stack.

For example we might have:
```ruby
	- initializers/money.rb
	Money.rounding_mode = BigDecimal::ROUND_HALF_EVEN

	- controllers/test.rb
	Money.with_rounding_mode(BigDecimal::ROUND_UP) do
	  object.amount = Money.from_amount(1.234)
	  object.save
	  other_object.amount = object.amount * different_rate
	end

	- models/object.rb
	def save
	  Money.with_rounding_mode(BigDecimal::ROUND_DOWN) do
	    self.tax = object.amount * tax_rate
	  end
	  ...
	end
```

Above example prior the change would cause other_object to use
`ROUND_HALF_EVEN`, despite it's explicitly defined in a `with_rounding_mode`
block which sets `ROUND_UP`.

The small rounding errors might be tricky to debug,
and in worst case scenarios can cumulate to a bigger differences when
code operates on multiple amounts that might be incorrectly rounded.

Co-Authored-By: Łukasz Wójcik <lukasz.wojcik@simplybusiness.co.uk>